requirements.txt
beautifulsoup4==4.14.3
blinker==1.9.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
cloudscraper==1.2.71
Flask==3.1.2
Flask-SQLAlchemy==3.1.1
greenlet==3.3.0
idna==3.11
instaloader==4.15
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
numpy==2.3.5
pandas==2.3.3
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.5
requests-toolbelt==1.0.0
six==1.17.0
soupsieve==2.8
SQLAlchemy==2.0.45
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.6.1
Werkzeug==3.1.4
